Split overview tab file
